third_party/WebKit/LayoutTests/fast/block/marquee-width-shrinks-to-fit-in-fixed-size-container.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<style>
.outer {
  width: 200px;
}
.inner {
  width: fit-content;
}
</style>
<div class="outer">
  <div class="inner">
    <marquee>The quick brown fox jumps over the lazy dog. The quick brown fox jumps over the lazy dog. The quick brown fox jumps over the lazy dog.</marquee>
  </div>
</div>
<script>
description("Test that a horizontal marquee doesn't contribute to a container's intrinsic minimum width.");
shouldBeEqualToNumber("document.querySelector('.inner').getBoundingClientRect().width", 200);
</script>
