third_party/blink/web_tests/platform/generic/fast/block/marquee-width-shrinks-to-fit-in-fixed-size-container-expected.txt

Test that a horizontal marquee doesn't contribute to a container's intrinsic minimum width.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.querySelector('.inner').getBoundingClientRect().width is 200
PASS successfullyParsed is true

TEST COMPLETE

The quick brown fox jumps over the lazy dog. The quick brown fox jumps over the lazy dog. The quick brown fox jumps over the lazy dog.
